Moves errors and post handlers to own files
